requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-google-genai>=0.0.5
langchain-community
faiss-cpu
transformers
sentence-transformers
google-generativeai
pymupdf
streamlit
python-dotenv
pytesseract
pdf2image
pillow